009 Define playing field.
Introduced the fixed BOARD_SIZE as the new axiom to calculate sizes from: a board consists of a fixed number of tiles. This board fills the available canvas space as much as possible.
Elements are created by this new board, severing the link between canvas (i.e. pixels) and elements (i.e. pieces on the board).